pyproject.toml
[tool.poetry]
name = "stravaweb"
version = "0.1.0"
description = ""
authors = ["Roman Haritonov <reclosedev@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7"
requests = "^2.22"
pyyaml = "^5.1"
requests-cache = "^0.5.2"
python-dotenv = "^0.13.0"
click = "^7.1.2"
lxml = "^4.5.2"
pandas = "^1.0.5"
numpy = "^1.19.0"
geopandas = "^0.8.1"
jupyterlab = "^2.2.0"
matplotlib = "^3.2.2"
folium = {git = "https://github.com/python-visualization/folium", rev = "master"}
branca = {git = "https://github.com/python-visualization/branca", rev = "master"}
dataset = "^1.3.2"
cssselect = "^1.1.0"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"

